chore: move search.py
